Share the network observable to avoid hitting the network twice
While the network operation is ongoing, another subscriber could come
and subscribe to the source. At that point, memory is empty and disk
as well because they're still waiting for the original network request
to come back and populate them. By sharing the observable we avoid
the scenario of multiple network requests while we still wait for the
response of one.